:bug: #30 번역 기능 모듈화 및 예외처리
